Move solar_secrets to own module
