--- common/src/solar_secrets/secrets.py
@@ -1,16 +1,0 @@
-import json
-from enum import Enum
-
-
-class SecretKey(str, Enum):
-    apiKey = "apiKey"
-    dnsPassword = "dnsPassword"
-    appid = "appid"
-
-
-def getSecret(secretKey: SecretKey):
-    secretsFilepath = f"/local/secrets.json"
-    with open(secretsFilepath, "r") as secretsFile:
-        secrets = json.load(secrets)
-
-    return secrets[secretKey]
